Rename package by removing birhanu.
