+++ app/src/androidTest/java/com/bizdesc/birhanu/activity/ApplicationTest.java
@@ -1,4 +1,4 @@
-package com.bizdesc.birhanu.myapplication;
+package com.bizdesc.birhanu.activity;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
