+++ app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.bizdesc.birhanu.myapplication"
+<manifest package="com.bizdesc.birhanu.activity"
           xmlns:android="http://schemas.android.com/apk/res/android">
 
   <uses-permission android:name="android.permission.INTERNET" />
@@ -11,7 +11,7 @@
     android:supportsRtl="true"
     android:theme="@style/AppTheme">
     <activity
-      android:name=".LoginActivity"
+      android:name="com.bizdesc.birhanu.activity.LoginActivity"
       android:label="@string/app_name"
       android:theme="@style/AppTheme.NoActionBar">
       <intent-filter>
@@ -29,13 +29,13 @@
       android:value="@integer/google_play_services_version"/>
 
     <activity
-      android:name=".CardActivity"
+      android:name="com.bizdesc.birhanu.activity.CardActivity"
       android:label="@string/title_activity_display_message"
-      android:parentActivityName=".LoginActivity"
+      android:parentActivityName="com.bizdesc.birhanu.activity.LoginActivity"
       android:theme="@style/AppTheme.NoActionBar">
       <meta-data
         android:name="android.support.PARENT_ACTIVITY"
-        android:value=".LoginActivity"/>
+        android:value="com.bizdesc.birhanu.activity.LoginActivity"/>
     </activity>
   </application>
